requirements.txt
python-telegram-bot==21.6
Flask==2.3.2
Werkzeug==2.3.6
SQLAlchemy==1.4.27
APScheduler==3.10.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0